Interconvert Strings and Integers - First Commit
A string is a sequence of characters. A string may encode an integer e.g., "123" encodes 123. In this problem you are to implement methods that take a string representing an integer and return the corresponding integer and vice versa. Your code could handle negative integers. You cannot use library functions like stoi in C++ and parseInt in java.