Test if a binary tree is Height Balanced - First Commit
A binary tree is sais to be height balanced if for each node in the tree, the difference in the height of its left and right subtrees is atmost one. A perfect binary tree is height balanced as is a complete binary tree. A height balanced binary tree does not have to be perfect or complete. Write a program that takes as input the root of a binary tree and checks whether the tree is height balanced.