Towers of Hanoi Problem - First Commit
Write a program which prints a sequence o operations that transfers n rings from one peg to another. You have a third peg which is initially empty. The only operation you cna perofrm is taking a single ring from the top of onoe peg and placing it on the top of another peg. You must never place a larger ring above a smaller ring.